backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
pandas==1.5.3
numpy==1.24.2
python-dotenv==1.0.0
